Adapt geogig plugin to upstream package refactoring
