+++ src/community/geogig/src/main/java/org/geogig/geoserver/config/LogStore.java
@@ -41,6 +41,11 @@ import org.springframework.beans.factory.InitializingBean;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 
+import com.google.common.base.Preconditions;
+import com.google.common.base.Splitter;
+import com.google.common.base.Strings;
+import com.google.common.base.Throwables;
+
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
@@ -48,11 +53,6 @@ import ch.qos.logback.classic.db.DBAppender;
 import ch.qos.logback.classic.encoder.PatternLayoutEncoder;
 import ch.qos.logback.core.db.DataSourceConnectionSource;
 import ch.qos.logback.core.util.StatusPrinter;
-
-import com.google.common.base.Preconditions;
-import com.google.common.base.Splitter;
-import com.google.common.base.Strings;
-import com.google.common.base.Throwables;
 
 public class LogStore implements GeoServerLifecycleHandler, InitializingBean {
