+++ src/community/geogig/src/main/java/org/geogig/geoserver/gwc/MinimalDiffBoundsConsumer.java
@@ -12,11 +12,11 @@ import java.util.List;
 
 import org.eclipse.jdt.annotation.Nullable;
 import org.geotools.geometry.jts.JTS;
-import org.locationtech.geogig.api.Bounded;
-import org.locationtech.geogig.api.Bucket;
-import org.locationtech.geogig.api.NodeRef;
-import org.locationtech.geogig.api.plumbing.diff.PreOrderDiffWalk;
-import org.locationtech.geogig.api.plumbing.diff.PreOrderDiffWalk.BucketIndex;
+import org.locationtech.geogig.model.Bounded;
+import org.locationtech.geogig.model.Bucket;
+import org.locationtech.geogig.model.NodeRef;
+import org.locationtech.geogig.plumbing.diff.PreOrderDiffWalk;
+import org.locationtech.geogig.plumbing.diff.PreOrderDiffWalk.BucketIndex;
 
 import com.google.common.base.Predicate;
 import com.google.common.base.Predicates;
