+++ src/community/geogig/src/main/java/org/geogig/geoserver/rest/InitRequestHandler.java
@@ -16,7 +16,7 @@ import org.geogig.geoserver.config.PostgresConfigBean;
 import org.geogig.geoserver.config.RepositoryManager;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.locationtech.geogig.api.GeoGIG;
+import org.locationtech.geogig.repository.GeoGIG;
 import org.locationtech.geogig.repository.Hints;
 import org.locationtech.geogig.rest.RestletException;
 import org.restlet.data.Form;
