+++ src/community/geogig/src/main/java/org/geogig/geoserver/security/NetworkSecurityHook.java
@@ -15,14 +15,14 @@ import java.util.regex.Pattern;
 import org.geogig.geoserver.config.ConfigStore;
 import org.geogig.geoserver.config.WhitelistRule;
 import org.geoserver.platform.GeoServerExtensions;
-import org.locationtech.geogig.api.AbstractGeoGigOp;
-import org.locationtech.geogig.api.Remote;
-import org.locationtech.geogig.api.hooks.CannotRunGeogigOperationException;
-import org.locationtech.geogig.api.hooks.CommandHook;
-import org.locationtech.geogig.api.plumbing.LsRemote;
-import org.locationtech.geogig.api.porcelain.CloneOp;
-import org.locationtech.geogig.api.porcelain.FetchOp;
-import org.locationtech.geogig.api.porcelain.PushOp;
+import org.locationtech.geogig.hooks.CannotRunGeogigOperationException;
+import org.locationtech.geogig.hooks.CommandHook;
+import org.locationtech.geogig.plumbing.LsRemote;
+import org.locationtech.geogig.porcelain.CloneOp;
+import org.locationtech.geogig.porcelain.FetchOp;
+import org.locationtech.geogig.porcelain.PushOp;
+import org.locationtech.geogig.repository.AbstractGeoGigOp;
+import org.locationtech.geogig.repository.Remote;
 import org.springframework.security.web.util.matcher.IpAddressMatcher;
 
 import com.google.common.base.Optional;
