+++ src/community/geogig/src/main/java/org/geogig/geoserver/web/repository/RepositoryEditFormPanel.java
@@ -25,14 +25,14 @@ import org.apache.wicket.model.Model;
 import org.apache.wicket.model.ResourceModel;
 import org.geogig.geoserver.config.RepositoryInfo;
 import org.geogig.geoserver.config.RepositoryManager;
-import org.locationtech.geogig.api.Context;
-import org.locationtech.geogig.api.GeoGIG;
-import org.locationtech.geogig.api.GeogigTransaction;
-import org.locationtech.geogig.api.Remote;
-import org.locationtech.geogig.api.plumbing.TransactionBegin;
-import org.locationtech.geogig.api.porcelain.RemoteAddOp;
-import org.locationtech.geogig.api.porcelain.RemoteListOp;
-import org.locationtech.geogig.api.porcelain.RemoteRemoveOp;
+import org.locationtech.geogig.plumbing.TransactionBegin;
+import org.locationtech.geogig.porcelain.RemoteAddOp;
+import org.locationtech.geogig.porcelain.RemoteListOp;
+import org.locationtech.geogig.porcelain.RemoteRemoveOp;
+import org.locationtech.geogig.repository.Context;
+import org.locationtech.geogig.repository.GeoGIG;
+import org.locationtech.geogig.repository.GeogigTransaction;
+import org.locationtech.geogig.repository.Remote;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
