+++ src/community/geogig/src/main/java/org/geogig/geoserver/web/security/SecurityLogsPanel.java
@@ -10,7 +10,6 @@ import java.util.List;
 
 import org.apache.wicket.AttributeModifier;
 import org.apache.wicket.Component;
-import org.apache.wicket.request.resource.PackageResourceReference;
 import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.extensions.ajax.markup.html.modal.ModalWindow;
 import org.apache.wicket.markup.html.basic.Label;
@@ -18,6 +17,7 @@ import org.apache.wicket.markup.html.basic.MultiLineLabel;
 import org.apache.wicket.markup.html.panel.Panel;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.model.Model;
+import org.apache.wicket.request.resource.PackageResourceReference;
 import org.apache.wicket.util.file.File;
 import org.geogig.geoserver.config.LogEvent;
 import org.geogig.geoserver.config.LogEvent.Severity;
