+++ src/community/geogig/src/main/resources/META-INF/services/org.locationtech.geogig.hooks.CommandHook
