+++ src/community/geogig/src/test/java/org/geogig/geoserver/functional/GeoServerFunctionalTestInjectorSource.java
@@ -7,6 +7,7 @@ package org.geogig.geoserver.functional;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Stage;
+
 import cucumber.api.guice.CucumberModules;
 import cucumber.runtime.java.guice.InjectorSource;
