+++ src/community/geogig/src/test/java/org/geogig/geoserver/rest/GeoGigWebAPIIntegrationTest.java
@@ -4,6 +4,7 @@
  */
 package org.geogig.geoserver.rest;
 
+import static org.junit.Assert.*;
 import static org.custommonkey.xmlunit.XMLAssert.assertXpathEvaluatesTo;
 import static org.custommonkey.xmlunit.XMLAssert.assertXpathExists;
 import static org.custommonkey.xmlunit.XMLAssert.assertXpathNotExists;
@@ -16,6 +17,7 @@ import java.io.EOFException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Random;
 
 import org.geogig.geoserver.GeoGigTestData;
@@ -34,15 +36,15 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.locationtech.geogig.api.GeoGIG;
-import org.locationtech.geogig.api.ObjectId;
-import org.locationtech.geogig.api.Ref;
-import org.locationtech.geogig.api.RevObject;
-import org.locationtech.geogig.api.plumbing.RefParse;
-import org.locationtech.geogig.api.plumbing.ResolveTreeish;
-import org.locationtech.geogig.api.plumbing.RevObjectParse;
 import org.locationtech.geogig.geotools.data.GeoGigDataStore;
 import org.locationtech.geogig.geotools.data.GeoGigDataStoreFactory;
+import org.locationtech.geogig.model.ObjectId;
+import org.locationtech.geogig.model.Ref;
+import org.locationtech.geogig.model.RevObject;
+import org.locationtech.geogig.plumbing.RefParse;
+import org.locationtech.geogig.plumbing.ResolveTreeish;
+import org.locationtech.geogig.plumbing.RevObjectParse;
+import org.locationtech.geogig.repository.GeoGIG;
 import org.locationtech.geogig.storage.ObjectSerializingFactory;
 import org.locationtech.geogig.storage.datastream.DataStreamSerializationFactoryV1;
 import org.opengis.feature.Feature;
@@ -54,6 +56,7 @@ import org.w3c.dom.Document;
 import com.google.common.base.Throwables;
 import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.Iterators;
+import com.google.common.collect.Lists;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
@@ -104,7 +107,6 @@ public class GeoGigWebAPIIntegrationTest extends GeoServerSystemTestSupport {
 
         geogigData.add().commit("Added test features");
 
-
         CatalogBuilder catalogBuilder = geogigData.newCatalogBuilder(catalog);
         int i = rnd.nextInt();
         catalogBuilder.namespace("geogig.org/" + i).workspace("geogigws" + i)
@@ -135,10 +137,10 @@ public class GeoGigWebAPIIntegrationTest extends GeoServerSystemTestSupport {
         assertNotNull(repositoryInfo);
         BASE_URL = "/geogig/repos/testrepo";
     }
-   
+
     @After
     public void after() {
-		RepositoryManager.close();
+        RepositoryManager.close();
     }
 
     /**
@@ -257,8 +259,10 @@ public class GeoGigWebAPIIntegrationTest extends GeoServerSystemTestSupport {
 
         ObjectSerializingFactory factory = DataStreamSerializationFactoryV1.INSTANCE;
 
-        Iterator<RevObject> objects = new ObjectStreamIterator(responseStream, factory);
-        RevObject actual = Iterators.getLast(objects);
+        List<RevObject> objects = Lists
+                .newArrayList(new ObjectStreamIterator(responseStream, factory));
+        assertFalse(objects.isEmpty());
+        RevObject actual = objects.get(objects.size() - 1);
         assertEquals(expected, actual);
     }
