+++ src/community/geogig/src/test/java/org/geogig/geoserver/security/SecurityLoggerTest.java
@@ -13,18 +13,18 @@ import static org.mockito.Mockito.verify;
 import org.geogig.geoserver.config.LogStore;
 import org.junit.Before;
 import org.junit.Test;
-import org.locationtech.geogig.api.Ref;
-import org.locationtech.geogig.api.Remote;
-import org.locationtech.geogig.api.plumbing.RefParse;
-import org.locationtech.geogig.api.porcelain.BlameOp;
-import org.locationtech.geogig.api.porcelain.CleanOp;
-import org.locationtech.geogig.api.porcelain.CloneOp;
-import org.locationtech.geogig.api.porcelain.DiffOp;
-import org.locationtech.geogig.api.porcelain.FetchOp;
-import org.locationtech.geogig.api.porcelain.PullOp;
-import org.locationtech.geogig.api.porcelain.PushOp;
-import org.locationtech.geogig.api.porcelain.RemoteAddOp;
-import org.locationtech.geogig.api.porcelain.RemoteRemoveOp;
+import org.locationtech.geogig.model.Ref;
+import org.locationtech.geogig.plumbing.RefParse;
+import org.locationtech.geogig.porcelain.BlameOp;
+import org.locationtech.geogig.porcelain.CleanOp;
+import org.locationtech.geogig.porcelain.CloneOp;
+import org.locationtech.geogig.porcelain.DiffOp;
+import org.locationtech.geogig.porcelain.FetchOp;
+import org.locationtech.geogig.porcelain.PullOp;
+import org.locationtech.geogig.porcelain.PushOp;
+import org.locationtech.geogig.porcelain.RemoteAddOp;
+import org.locationtech.geogig.porcelain.RemoteRemoveOp;
+import org.locationtech.geogig.repository.Remote;
 import org.mockito.ArgumentCaptor;
 
 public class SecurityLoggerTest {
