+++ src/community/geogig/src/test/java/org/geogig/geoserver/web/repository/CommonPanelTest.java
@@ -4,7 +4,6 @@
  */
 package org.geogig.geoserver.web.repository;
 
-import static org.geoserver.web.GeoServerWicketTestSupport.tester;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
